gadgets/bpfstats: Do not require --enable-bpfstats
The current implementation of the bpfstats gadget requires the daemon
to be deployed with the --enable-bpfstats parameter to enable eBPF stats
collection in the kernel. This UX is bad as users will need to enable it
from the beginning, having a performance issue for all gadgets that are
run, and by not being able to use this functionality on an existing IG
deployment without having to restart it.

This PR changes the approach to enable the stats only when the bpfstats
gadget is run, which is more practical as users don't need to change
any configuration at deploy time and the stats collection is only enabled
while the bpfstats gadget runs.

By changing this, we also fix an issue where stats collection wasn't
disabled after running Inspektor Gadget.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>