ci: Enable cache for building helper images
We aren't caching the Docker layers for helper images like the
gadget-builder. It causes the following:
1. The building takes very long (~5mins)
2. A new built image has a totally different set of digest for its layers
compared to one generated in a previous builder as our images don't
have "reproducible builds" support.

The second point affects users compiling gadgets, as each time the image
was changed upstream, they have to pull all the layers (hundreds of MBs).

This commit enables the cache fixing the issues described above. Now,
also the logic to avoid building the image based on some file patterns
is removed as we can always try to build it with a very minimum overhead.

Suggested-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>
Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>