user_stack_map.h: fix compilation warnings
We have the following code pattern:

	// vmlinux.h
	struct pid {
	        struct upid numbers[1];
	};
	// include/gadget/user_stack_map.h
	struct upid *numbers = &thread_pid->numbers;

It generates the following compilation warning:

> /work/include/gadget/user_stack_map.h:148:15: warning: incompatible
> pointer types initializing 'struct upid *' with an expression of type
> 'struct upid (*)[1]' [-Wincompatible-pointer-types]

We cannot solve it with the following code:
	struct upid *numbers = &thread_pid->numbers[0];

because Linux v6.5 commit b69f0aeb0689 ("pid: Replace struct pid
1-element array with flex-array") made the following change:

-       struct upid numbers[1];
+       struct upid numbers[];

And it would generate a CORE relocation with an accessor for the first
entry of the array. The relocation would fail at load time because
flex-array are zero-sized in BTF.

So we unfortunately have to cast the pointer to make it work without
warnings both on Linux >= v6.5 and < v6.5.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>