pkg/operators/ebpf: add support for trailing data
This allows adding a trailing (rest) data field to a datasource that
from annotations. It will be populated with data that exceeds the length
of the expected BTF type. It can optionally request the length of
the trailing data from an existing uint64 field.

WIP. Currently only works for ring buffers.

Signed-off-by: Michael Friese <mfriese@microsoft.com>