oci: Avoid using internet to get gadget signature
The previous logic always tried to use internet to download the gadget
signature from the container registry. This approach made it impossible
to run a gadget in a system without connectivity, even if the image
has been pulled before. This commit changes the logic to cache the
signature when pulling the image, so it's available even when there's
not connectivity.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>