socketenricher: Use CO-RE to make it resilient against changes
The socketenricher operator is used  to enrich events with process
information from a socket. This is implemented as an ebpf map that
stores a list of tuples (IPs+ports) and the information of the process
that created the socket. This map is an API contract between the gadgets
and Inspektor Gadget, any change to the layout of the data on it will
require the gadgets to be recompiled. Of course, this is a very big
limitation.

This commit uses the existing CO-RE technology to make it possible to
update the layout of this map without breaking the gadgets. This is done
by generating a BTF data that represents the structure stored on this map,
then this data is fed to the ebpf-go library when loading the program,
ensuring it relocates the access to that structure.

This in practice allows to:
- Introduce new information to the socket enricher
- Change the type or position of existing fields on it
- Dynamically disable or reduce the maximum size of some expensive fields (like paths).
  - Reduce the size of the map
  - Avoid spending CPU cycles on disabled fields

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>