operators/ebpf: Use a BTF-based solution to register gadgets
Before the patch the GADGET_TRACER(), GADGET_MAP_ITER(), etc. macros
were used to define gadgets in the ebpf code. Those macros internally
created a var starting with a given prefix and used the var name to
store all the information about the gadget: name, types, etc.
Relying on string concatenation has some issues, for instance, extending
that to support more information in the future is difficult.

This commit uses an approach similar to the one used to define eBPF maps:
defining a structure that generates BTF information. It has the following
advantages:
- Easier to extend later on, as adding more information is just a matter
of adding fields to the struct
- Uses a well-known approach to define metadata in the program (as used for BTF maps)

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>