Add logs for gadget start
Some times running a gadget takes a bit long: while the image is pulled,
wasm and ebpf programs are loaded, etc. This commit introduces a mechanism
to inform the user about the different steps when running a gadget.

The logging always use text, json support will be added later on.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>