oci: Use oras to cache gadget signature
Gadget signatures are pushed to the registry as OCI artifacts. Use the
oras logic to pull them to the local OCI store instead of using a custom
logic.

This also makes it easier to import and export the signatures as they
are handled as standard oci artifacts, so we don't need any special logic.
(This will be done in a future commit)

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>