ci: Revome usage of ttl.sh
The CI uses ttl.sh as a "junk" repository to store gadget images.
Unfortunately, it fails some times, so this commit switches to use ghcr.io
instead. As ghcr.io isn't designed for that purpose, we need to use the
tag to store multiple gadgets (even from different runs) on the same repo.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>