kallsyms/cache: preload kallsyms that cilium/ebpf will need
Introduce a new package pkg/kallsyms/cache used by:
- kfilefield
- container-hook
- kfilefields

The cost of loading many symbols at once is lower than loading few
symbols multiple times for multiple ebpf collections. This is because
most of the cpu time is used for reading /proc/kallsyms in the kernel
function module_get_kallsym. So it is better to read /proc/kallsyms only
one time.
https://github.com/torvalds/linux/blob/v6.15/kernel/module/kallsyms.c#L376

The cache package provides two separate functions:
- RegisterSymbolsFromSpec: to be called from init()
- PopulateKallsymsCache: to be called just before loading your ebpf
  collection

In this way, different users of the cache don't have to collaborate to
collect the symbols they use.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>