ebpf: Support configuring length of big fields for gadgets
Some gadgets provide some fields that can be quite big (cwd, exepath, etc).
In some scenarios it's needed to change their maximum size to save memory.
This commit completes the work initiated in a9ceda8 ("gadgets/trace_dns: Use optional fields on socket enricher")
by allowing users to set the max size of expensive fields in the event
structure. This is done by creating a new BTF spec with the modified
structures and passing it as a target relocation.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>