ebpf: Support configuring length of big fields for gadgets
Some gadgets provide some fields that can be quite big (cwd, exepath, etc).
In some scenarios it's needed to change their maximum size to save memory.
This commit completes the work initiated in  ("gadgets/trace_dns: Use optional fields on socket enricher")
by setting the size of the fields in the event structure according to the
configuration of the socker enricher. This is done by creating a new BTF
spec with the modified structures and passing it as a target relocation.

Support for this in all gadgets (not only fields from the socker enricher)
will be added in another commit later on.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>