cmd/common/image/import: Support JSON output mode
Considering the output of `ig image import` command is a list of images
imported from the specified file, it makes sense to support JSON output
mode. This allows users to easily parse the output programmatically,
especially when integrating with other tools or for scripting purposes
(e.g., in tests).

Maintains backward compatibility with the existing human-readable output
as the default.

Signed-off-by: Jose Blanquicet <josebl@microsoft.com>