in-tree gadget build: use intree vmlinux
When building in-tree gadgets, the $IG_SOURCE_PATH environment variable
is set to use the header files from the IG sources rather than the
builder container. This was done correctly for the headers in
include/gadget/ such as mntns.h, but not for vmlinux.h.

This patch adds support for using vmlinux.h from the IG sources.

The problem was found out after updating vmlinux.h: the new types were
not available for in-tree gadgets.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>