tools: Add bench tool
The bench tool is used as a load generator in order to measure the
performance overhead of Inspektor Gadget.

The tool supports different kind of events:
- dns
- open
- exec
- file-io
- processes
- tcp

It tries to generate a fixed rate of events per second that can be
configured by the user.

It also includes two helpers to act as server for networking loads
- dns-server
- tcp-server

TODO:
- build and push the container image from the CI
- code review and clean up
  - split generators on different packages?
  - is numer of goroutines used the right one? Should it be configurable?
- How generators could expose custom flags?
  - processes shoulnd't use events-per-second
- Documentation

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>
Signed-off-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>