ebpf operator: automatically change map size at runtime
For some gadgets, it can be useful to have a parameter to change the
size of a map on the command line.

For example, a gadget with this map:

struct {
       __uint(type, BPF_MAP_TYPE_HASH);
       __uint(key_size, sizeof(u32));
       __uint(value_size, sizeof(u32));
       __uint(max_entries, 0);
} mymap SEC(".maps");

It could have the following parameter, named ${MAP_NAME}_max_entries:

const volatile int mymap_max_entries = 5;
GADGET_PARAM(mymap_max_entries);

And the following in gadget.yaml:

params:
  ebpf:
    mymap_max_entries:
      key: max-entries
      defaultValue: "5"

With this patch, the ebpf operator will automatically change the
map's max_entries.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>