op/otel-metrics: Add matricname annotation
A datasource can have multiple fields with the same name, if they
belong, for example, to different structs inside the datasource. top_tcp
has a src.k8s.name and dst.k8s.name and both fields keys would be
"name". To differentiate between this the metricname annotation can be used
to override this default behavior

Signed-off-by: Burak Ok <burakok@microsoft.com>