github actions: disable some cron jobs on forks
Before this patch, forks would run daily github actions to check the
links, check the artifact hub tags and run the scorecard analysis.

Most of the time, forks are just there to make pull requests. The main
branch is unused and becomes stale. It is a waste of resources to run
daily checks on it.

This patch disables these 3 cron jobs. Other cron jobs are left
unchanged because they are already disabled by default (unless the fork
configures environment variables to enable them).

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>