ustack: refactor to build on different platform
Before this patch, the ustack package only builds on Linux. In
preparation for making the ustack operator work client-side, this patch
moves part of the code in _linux.go files. It will make it easier for
future patches.

Tested by adding the import in a cmd package:
+++ b/cmd/common/oci.go
+       _ "github.com/inspektor-gadget/inspektor-gadget/pkg/operators/ustack"

And running "make gadgetctl-all"

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>