ustack: use operator lifecycle correctly
According to the operator lifecycle documentation:
- InstantiateDataOperator: creates datasources. Called both in
  GetGadgetInfo and RunGadget
- PreStart: the operator subscribes to datasources
- Start: the operator starts emitting data (nothing to do in the case of
  the ustack operator)

This patch moves datasource subscription from Start to PreStart. This
ensure that ustack enrichment happens from the very first events.

This patch also instantiates the symbolizer in PreStart instead of
InstantiateDataOperator. The symbolizer is not needed for GetGadgetInfo.
The symbtab symbolizer is also not used client-side.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>