ustack: cache the runtime base addresses
Due to Address Space Layout Randomization (ASLR), most processes have a
random base address chosen at execve-time.

Before this patch, IG parsed /proc/pid/maps to find the runtime base
address at each stack trace resolution. This patch adds a cache to
minimize the overhead.

The cache is not indexed by pid because:
- The base address changes with execve but the pid does not change at
  that time
- Several pids can share the same memory space with CLONE_VM (e.g.
  threads)

Instead, I use the mm_struct address to distinguish the programs sharing
the same base address. There are however a possibility where the
mm_struct address is reused: if the mm_struct is freed and re-allocated
to the same address. To avoid this scenario, I add mm->start_code and
mm->start_stack in the hash. A new program reusing the same mm_struct
address will have a different random start_code and start_stack.

To reduce the size of events sent to the ebpf ring buffer, we send a u32
hash:

```c
	// Identify the base address.
	// Changes whenever the base address changes at execve.
	// Not shared between processes because of Address Space Layout Randomization.
	// FNV-1a hash of tuple (mm_addr, start_code, start_stack)
	__u32 base_addr_hash;
```

I chose the hash algorithm FNV-1a because it is fast and easy to
implement in ebpf. It is not cryptographically secure but the 3
components of the tuple (mm_addr, start_code, start_stack) are chosen by
the kernel.

Note that the base address cannot reliably be found from ebpf. Even
though mm_struct has fields like start_code and start_data, that does
not directly give the base address. The information could be found in
kernel space by iterating over the VMAs but this cannot be done in ebpf
because that would be an unbounded loop.

Example:
```
$ sudo cat /proc/1853244/maps
55f86c7ec000-55f86c7ed000 r--p 00000000 08:16 23894993                   /usr/local/bin/python3.13
55f86c7ed000-55f86c7ee000 r-xp 00001000 08:16 23894993                   /usr/local/bin/python3.13
55f86c7ee000-55f86c7ef000 r--p 00002000 08:16 23894993                   /usr/local/bin/python3.13
55f86c7ef000-55f86c7f0000 r--p 00002000 08:16 23894993                   /usr/local/bin/python3.13
55f86c7f0000-55f86c7f1000 rw-p 00003000 08:16 23894993                   /usr/local/bin/python3.13
...
55f86d141000-55f86d704000 rw-p 00000000 00:00 0                          [heap]
...
7ffd65563000-7ffd65584000 rw-p 00000000 00:00 0                          [stack]
```

mm_struct fields (found by modifying the snapshot_process gadget):

| Field Name       | Address (Hexadecimal) |
| :---             | :---                  |
| **start_code**   | 55f86c7ed000          |
| **end_code**     | 55f86c7ed155          |
| **start_data**   | 55f86c7efdb0          |
| **end_data**     | 55f86c7f0018          |
| **start_brk**    | 55f86d141000          |
| **start_stack**  | 7ffd65583590          |

The base address is 55f86c7ec000 and the start_code is 55f86c7ed000
(8KiB below). While it is often 8KiB in practice, we cannot rely on
that. It is 8KiB here because the first LOAD segment has a size of 8KiB
(after rounding up on the page size). It typically contains metadata
like the ELF header, the program headers and the interpreter path
("/lib64/ld-linux-x86-64.so.2"). Other programs might be compiled
differently.

```
$ sudo readelf -lW /usr/local/bin/python3.13
Program Headers:
  Type           Offset   VirtAddr           PhysAddr           FileSiz  MemSiz   Flg Align
  PHDR           0x000040 0x0000000000000040 0x0000000000000040 0x000310 0x000310 R   0x8
  INTERP         0x000394 0x0000000000000394 0x0000000000000394 0x00001c 0x00001c R   0x1
      [Requesting program interpreter: /lib64/ld-linux-x86-64.so.2]
  LOAD           0x000000 0x0000000000000000 0x0000000000000000 0x000760 0x000760 R   0x1000
  LOAD           0x001000 0x0000000000001000 0x0000000000001000 0x000155 0x000155 R E 0x1000
  LOAD           0x002000 0x0000000000002000 0x0000000000002000 0x0000f0 0x0000f0 R   0x1000
  LOAD           0x002db0 0x0000000000003db0 0x0000000000003db0 0x000268 0x000270 RW  0x1000
...
```

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>