enrich: Do not try to enrich multiple times
Inspektor Gadget is able to enrich events by using the mount namespace
or the network namespace. The previous implementation had an issue as
it first tried to enrich the mount namespace, and then with the network
namespace. This caused the specific information provided by the mount ns
(like the container in a pod) to be lost as it was overwritten by the
net ns enrichment. (the network namespace is shared by all containers
on the same pod so it's not possible to distinguish among them when
enrichment with this field). This commit fixes that logic by introducing
some logic that avoid to perform a second enrichment if the first one
succeded.

Fixes: b1875da7aed0 ("pkg/types: add wrapper for events to implement interfaces for compatibility")
Reported-by: Matthias Bertschy <matthias.bertschy@gmail.com>
Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>