Revert "ci: bump sigstore/cosign-installer from 3.10.0 to 4.0.0"
This reverts commit 2e1075568fb01b6b25541d397f81fd12edf7427f.
This commit broke the CI, as the gadgets are not signed but the failure is not
notified [1,2].

Let revert it to have more time to investigate the root cause and change the
invocation of signing.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://github.com/inspektor-gadget/inspektor-gadget/actions/runs/18678704075/job/53255123822#step:10:1
The whole text to answer "yes" to cosign pushing something is not printed with
the new version, thus the gadgets are not signed and CI fails later.
[2]: https://github.com/inspektor-gadget/inspektor-gadget/actions/runs/18596598596/job/53028651000#step:10:32
The whole text is printed, which ensures the gadget are signed.