operators/ebpf: forward lostSampleCount for PerfEventArray
This adds the number of lost samples between the current and the previous event for PerfEventArray.
Support for RingBuf is a bit more complicated and will be added in a follow-up PR. This does NOT
emit an event for every reported sample loss - instead, it accumulates the number and sends it with
the next successfully read event.

To make sure the number of lost samples is not lost with dropped packets (on the network layer), we
need to take this number into account when calculating the packet seq number.

Signed-off-by: Michael Friese <mfriese@microsoft.com>