grpc: carry lostSampleCount
This takes the lostSampleCount from individual packets and creates a monotonically
creating counter per DataSource from it. It removed the lostSampleCount from the
packet (to save some bytes over the wire) and puts the counter into the GadgetEvent
instead.

This way we don't lose the lostSampleCount in case a packet is lost; instead, the
client is able to detect lost packets and lost samples individually. On the client,
the per-packet information is restored to the number of lost samples relative to the
previous packet.

Signed-off-by: Michael Friese <mfriese@microsoft.com>