pkg: oci: Pull gadget signature if verification failed.
The gadget could be signed in a remote repository, but the signature may not be
in the local OCI store. This could occur if the image was build locally, pushed
to a remote repository and signed [1].
Let's first try to verify the gadget locally and pull the signature from remote
repository it if fails to then verify it again.

Reported-by: Mauricio Vásquez <mauriciov@microsoft.com>
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://github.com/inspektor-gadget/inspektor-gadget/pull/5086#pullrequestreview-3504937430